Crate tower_iter

Extensions to Tower aiming at ease of use for sets and iterators of Services.

This crate is not no_std compatible, as it requires tower::util.

Re-exports

pub use call_iter::MapM;

pub use call_iter::MapMExt;

pub use error::Error;

pub use future_ext::FutureExt;

pub use future_ext::MapOk;

pub use service_set::ServiceSet;

Modules

call_iter

A set of types dedicated to running multiple services.

error

future_ext

iter_svc

service_multi_iter

service_multiset

service_set