tower_iter

Module call_iter

A set of types dedicated to running multiple services.

Structs

MapM

MapReady

Converts an Iterator over a set of (S, I) where S is a service that takes the input I into an iterator over the futures that yield the service once it is ready.

MapServiceCall

Converts an Iterator over a set of (S, I) where S is a service that takes the input I into an iterator over the future from ServiceExt::oneshot.

ReverseTuple

Converts an iterator of (Left,Right) into an iterator of (Right,Left).

ServiceCall

Traits

MapMExt