tower_iter::call_iter

Trait MapMExt

pub trait MapMExt: Iterator + Sized {
    // Provided methods
    fn map_service_call<S, I>(self) -> MapServiceCall<Self, S, I>  { ... }
    fn reverse_tuple<I1, I2>(self) -> ReverseTuple<Self, I1, I2>  { ... }
}

Provided Methods

fn map_service_call<S, I>(self) -> MapServiceCall<Self, S, I>

Maps an iterator over (S, I), where S is a Service and I is the input to that service. Into an iterator over the results of calling each S with the given inputs.

use core::{
  convert::Infallible,
  iter::repeat_n,
};
use tower::service_fn;
use tower::Service;
use futures_lite::future::block_on;
use tower_iter::MapMExt;

async fn mul_2(i: u32) -> Result<u32, Infallible> {
  Ok(i * 2)
}
let mut mul_svc = service_fn(mul_2);
let mut iter = repeat_n(mul_svc, 5)
  .zip([
      5, 10, 15, 20, 25
  ].into_iter())
  .map_service_call();

assert_eq!(block_on(iter.next().unwrap()), Ok(10));
assert_eq!(block_on(iter.next().unwrap()), Ok(20));
assert_eq!(block_on(iter.next().unwrap()), Ok(30));
assert_eq!(block_on(iter.next().unwrap()), Ok(40));
assert_eq!(block_on(iter.next().unwrap()), Ok(50));
assert!(iter.next().is_none());

fn reverse_tuple<I1, I2>(self) -> ReverseTuple<Self, I1, I2>

Reverses a 2-tuple’s order.

use tower_iter::MapMExt;
let iter = &mut [
  (0, 1),
  (2, 3),
  (4, 5),
].into_iter().reverse_tuple();

assert_eq!(iter.next(), Some((1, 0)));
assert_eq!(iter.next(), Some((3, 2)));
assert_eq!(iter.next(), Some((5, 4)));
assert_eq!(iter.next(), None);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.